+++ OSBarcodeLib/Scanner/OSBARCScannerBehaviour.swift
@@ -1,3 +1,4 @@
+import AVFoundation
 import Combine
 import SwiftUI
 
@@ -5,6 +6,7 @@ import SwiftUI
 final class OSBARCScannerBehaviour: OSBARCCoordinatable, OSBARCScannerProtocol {
     /// A publisher value responsible for the resulting scanned value.
     @Published private var scanResult: String = ""
+    
     /// The publisher's cancellable instance collector.
     private var cancellables: Set<AnyCancellable> = []
     
@@ -26,7 +28,12 @@ final class OSBARCScannerBehaviour: OSBARCCoordinatable, OSBARCScannerProtocol {
                 self.scanResult = $0
             }
         )
-        let scannerView = OSBARCScannerView(scanResult: scanResultBinding)
+
+        // Get the default camera for capturing videos. This object will allows us to fetch the `hasTorch` method.
+        let captureDevice = AVCaptureDevice.default(.builtInWideAngleCamera, for: .video, position: .back)
+        let cameraHasTorch = captureDevice?.hasTorch ?? false
+        
+        let scannerView = OSBARCScannerView(scanResult: scanResultBinding, cameraHasTorch: cameraHasTorch, captureDevice: captureDevice)
         let hostingController = UIHostingController(rootView: scannerView)
         
         self.coordinator.present(hostingController)