+++ OSBarcodeLib/Scanner/OSBARCScannerViewControllerRepresentable.swift
@@ -1,13 +1,17 @@
+import AVFoundation
 import SwiftUI
 
 /// Structure responsible for bridging `OSBARCScannerViewController` into SwiftUI.
 struct OSBARCScannerViewControllerRepresentable: UIViewControllerRepresentable {
     /// The object containing the value to return.
     @Binding var result: String
+    /// The camera used to capture video for barcode scanning.
+    var captureDevice: AVCaptureDevice?
     
     func makeUIViewController(context: Context) -> OSBARCScannerViewController {
         let controller = OSBARCScannerViewController()
         controller.delegate = context.coordinator
+        controller.captureDevice = captureDevice
         
         return controller
     }