+++ docs/CHANGELOG.md
@@ -6,4 +6,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+- Add Torch Button feature (https://outsystemsrd.atlassian.net/browse/RMET-2749).
 - Implement the Scan Barcode feature (https://outsystemsrd.atlassian.net/browse/RMET-2748).